Linear Comments
This commit introduces linear comments. Right now the only thing you can
comment is a story. Comments on comments is not yet possible. Technically
the comments are modeled as EmbeddedDocuments.

The handing is done in the story model. Another interessting piece is the
routing. It uses nested routes to reflect that you cannot access a comment
directly. But nested routes add some pain in respect to forms an such
things.